water bottle which is kind of baloon
We all carry water bottles in our back pack. when it is full with water it is normal that it fills space in the bag. but when it is empty it holds unnecessary space in the bag. 
this is y I want a water bottle which is like a baloon, which can shrink into the size accordingly to the amount of water inside. the bottle must be reuseable, resisstant, flexible and easy to clean.